Add output suppression macro
Why:
Since the new strategy for deprecation changed from:
```ruby
  ActiveSupport::Deprecation.warn("message", attribute_caller)
```
to use a strategy:
```ruby
  ActiveSupport::Deprecation.new("6.0", "factory_bot")
```
The deprecations were sent directly to the default output and the flag
to silence them was not being take into consideration. Maybe it's worth
exploring opening a PR on the ActiveSupport side to include this
behavior.
In the meantime silencing the output is an option added to the
`:silence_deprecation` attribute on examples.

This commit:
- Adds the `SilenceOutput` module to suppress `stdout` and `stderr`
writes inside an example.
- Extracts block `with_silenced_output` to be used from the deprecation
macro.